update: Refactor form reset action
