+++ src/components/accessories/admin/exams/examForm/ExamForm.tsx
@@ -9,7 +9,6 @@ import React, {
   useEffect,
   useMemo,
   useRef,
-  useState,
 } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router";
@@ -34,7 +33,7 @@ import { IExamProps } from "./types";
 import { AddCircle, ChevronLeft, Delete } from "@mui/icons-material";
 import { IconButton, Radio, RadioGroup } from "@mui/material";
 import { useAppDispatch, useAppSelector } from "libraries/hooks/redux";
-import { useDiscardHelpers } from "libraries/hooks/ui";
+import { useDiscardHelpers, useResetFormHelpers } from "libraries/hooks/ui";
 import AutocompleteField from "../../../autocompleteField/AutocompleteField";
 import InfoBox from "../../../infoBox/InfoBox";
 
@@ -50,8 +49,6 @@ const ExamForm: FC<IExamProps> = ({
   const { t } = useTranslation();
   const navigate = useNavigate();
   const infoBoxRef = useRef<HTMLDivElement>(null);
-
-  const [openResetConfirmation, setOpenResetConfirmation] = useState(false);
 
   const examStore = useAppSelector((state) => state.exams);
 
@@ -150,17 +147,11 @@ const ExamForm: FC<IExamProps> = ({
       : "";
   };
 
-  const handleResetConfirmationDialog = useCallback(
-    (value: boolean) => () => {
-      setOpenResetConfirmation(value);
-    },
-    [setOpenResetConfirmation]
-  );
-
-  const handleResetConfirmation = useCallback(() => {
-    setOpenResetConfirmation(false);
-    formik.resetForm();
-  }, [navigate, formik.resetForm, setOpenResetConfirmation]);
+  const {
+    openResetConfirmation,
+    handleResetConfirmation,
+    handleResetConfirmationDialog,
+  } = useResetFormHelpers(formik as any);
 
   const addExamRow = useCallback(() => {
     formik.setFieldValue("rows", [...formik.values.rows, ""]);
