+++ src/components/accessories/admin/types/components/diseases/diseaseTypesForm/DiseaseTypeForm.tsx
@@ -1,16 +1,9 @@
 import { ChevronLeft } from "@mui/icons-material";
 import { useFormik } from "formik";
 import { useAppDispatch, useAppSelector } from "libraries/hooks/redux";
-import { useDiscardHelpers } from "libraries/hooks/ui";
+import { useDiscardHelpers, useResetFormHelpers } from "libraries/hooks/ui";
 import { get, has } from "lodash";
-import React, {
-  FC,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import React, { FC, useCallback, useEffect, useMemo, useRef } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router";
 import { object, string } from "yup";
@@ -44,8 +37,6 @@ const DiseaseTypeForm: FC<IDiseaseTypeFormProps> = ({
   const { t } = useTranslation();
   const navigate = useNavigate();
   const infoBoxRef = useRef<HTMLDivElement>(null);
-
-  const [openResetConfirmation, setOpenResetConfirmation] = useState(false);
 
   const diseaseTypesStore = useAppSelector((state) => state.types.diseases);
 
@@ -96,17 +87,11 @@ const DiseaseTypeForm: FC<IDiseaseTypeFormProps> = ({
       : "";
   };
 
-  const handleResetConfirmationDialog = useCallback(
-    (value: boolean) => () => {
-      setOpenResetConfirmation(value);
-    },
-    [setOpenResetConfirmation]
-  );
-
-  const handleResetConfirmation = useCallback(() => {
-    setOpenResetConfirmation(false);
-    formik.resetForm();
-  }, [navigate, formik.resetForm, setOpenResetConfirmation]);
+  const {
+    openResetConfirmation,
+    handleResetConfirmation,
+    handleResetConfirmationDialog,
+  } = useResetFormHelpers(formik as any);
 
   const cleanUp = useCallback(() => {
     if (creationMode) {
