+++ src/components/accessories/admin/vaccines/vaccineForm/VaccineForm.tsx
@@ -1,16 +1,9 @@
 import { ChevronLeft } from "@mui/icons-material";
 import { useFormik } from "formik";
 import { useAppDispatch, useAppSelector } from "libraries/hooks/redux";
-import { useDiscardHelpers } from "libraries/hooks/ui";
+import { useDiscardHelpers, useResetFormHelpers } from "libraries/hooks/ui";
 import { get, has } from "lodash";
-import React, {
-  FC,
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import React, { FC, useCallback, useEffect, useMemo, useRef } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router";
 import { object, string } from "yup";
@@ -47,8 +40,6 @@ const VaccineForm: FC<IVaccineFormProps> = ({
   const { t } = useTranslation();
   const navigate = useNavigate();
   const infoBoxRef = useRef<HTMLDivElement>(null);
-
-  const [openResetConfirmation, setOpenResetConfirmation] = useState(false);
 
   const vaccineStore = useAppSelector((state) => state.vaccines);
 
@@ -117,17 +108,11 @@ const VaccineForm: FC<IVaccineFormProps> = ({
       : "";
   };
 
-  const handleResetConfirmationDialog = useCallback(
-    (value: boolean) => () => {
-      setOpenResetConfirmation(value);
-    },
-    [setOpenResetConfirmation]
-  );
-
-  const handleResetConfirmation = useCallback(() => {
-    setOpenResetConfirmation(false);
-    formik.resetForm();
-  }, [navigate, formik.resetForm, setOpenResetConfirmation]);
+  const {
+    openResetConfirmation,
+    handleResetConfirmation,
+    handleResetConfirmationDialog,
+  } = useResetFormHelpers(formik as any);
 
   const onBlurCallback = useCallback(
     (fieldName: string) =>
